# perf: optimize HeadersList.get