--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -208,9 +208,7 @@ class HeadersList {
 
   // https://fetch.spec.whatwg.org/#concept-header-list-get
   get (name) {
-    name = name.toLowerCase()
-
-    const value = this[kHeadersMap].get(name)
+    const value = this[kHeadersMap].get(name.toLowerCase())
 
     // 1. If list does not contain name, then return null.
     // 2. Return the values of all headers in list whose name